* Updates to allow software control over choice of wifi connection method (either the software supplicant, or the broadcom firmware supplicant). These changes remove the "supplicant" compiler directive
* Fix bug causing supplicant to only make one attempt to reconnect if initial
connection attempt is not successful

* tidy up some error messages from the driver

* Update demo application to match the driver changes

* Remove the large IP address output as many people don't have screens big
enough to see it so it becomes unreadable.

* Remove unnecessary code for redirecting stdio in demo program.

* Fix missing parameter in supplicant ultibo layer.

* Add supplicant logging level control (see constants MSG_* and variable WPASupplicantLogLevel)



Note that for Pi400 and PiZero2W, the software supplicant MUST be used as
the broadcom firmware does not support a supplicant for those devices.

To choose the broadcom firmware supplicant, specify SUPPLICANT=FALSE in cmdline.txt
Any other value for SUPPLICANT= will make the application choose the software supplicant.

The {$ifdef supplicant} conditional define is now no longer required.

Note that when adding the driver to your own programs, in order to make a
connection, you must now call the appropriate connection function:

Call WirelessJoinNetWork() to join using the supplicant. Connection details will
be read by the supplicant from wpa_supplicant.conf
Call FirmwareWirelessJoinNetork() to join using the broadcom firmware supplicant.
Connection details must be supplied in the parameters. It's up to you
how you store them.

Why two options? Well, the intention is that the software supplicant be
used mostly, but there are still occasions where a device can't seem to
get connected via the software supplicant so the firmware supplicant
is there as a backup for those devices that still support it. Once we've
ironed out the remaining rare issues with connections then we won't need
the firmware supplicant anymore.